requirements.txt

apache-airflow>=2.8.0
apache-airflow-providers-postgres>=5.10.0
psycopg2-binary>=2.9.9